\name{GuessEncoding}
\alias{GuessEncoding}
\encoding{UTF-8}
\title{GuessEncoding}
\description{
  Guesses the encoding of a file.
}

\usage{
GuessEncoding(filename)
}
\arguments{
  \item{filename}{the name of the file to guess the encoding.}
}
\value{
  The name of the encoding of the file.
}

\author{
  Maëlick Claes <himself@maelick.net>
}

\keyword{file}
\keyword{guess}
\keyword{encoding}
